requirements.txt
aiohttp==3.7.3
alabaster==0.7.12
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
coverage==5.5
discord.py==1.6.0
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
packaging==20.9
pony==0.7.14
psycopg2==2.8.6
Pygments==2.8.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
snowballstemmer==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
